go logs-ingestion: Add example logs-ingestion tests
