+++ Makefile
@@ -13,3 +13,5 @@ restart:
 down:
 	docker compose down
 
+test:
+	docker compose -f docker-compose.test.yml up --build
