+++ client/Dockerfile
@@ -1,10 +1,7 @@
 # 1. For build React app
 FROM node:lts AS development
 
-ARG VITE_GOOGLE_CLIENT_ID
-ARG VITE_GOOGLE_REDIRECT_URI
 ARG VITE_BACKEND_URL
-ARG VITE_BACKEND_WS_URL
 
 WORKDIR /app
